Added + styled Hiragana Quiz/Test page
Added a few new hiragana quiz pages. They all stem from one component quiz page with two props passed (the characters to quiz and the link to the next lesson). The pages consist of a variable number of boxes, each containing a hiragana letter and an input box beneath. It is also fully resizable for mobile, with the background scaling properly.

Most of my time today was spent redesigning the page, and wrestling with the background image to get it to tile and place itself correctly without bugs. I did a lot of research in morphic designs, and I even created a full neumorphic quiz page in CSS, but I scrapped it because it was just too boring and didn't quite work with the amount of elements my page had. The current look aligns much more closely with the rest of the app's aesthetic. I had hoped to make the page fully functional by the end of the day, but I had to spend a lot of time revising the design. Currently, you can input text into each box, but there's no validation if any of the inputted text is correct or not (if it matches the hiragana character above).