Added several lessons. Imported and set up kana to hiragana conversion tool with type annotations and dictionary files in public folder.
I've been on vacation with family for a couple days, but I managed to get a little bit done. I've added at least half a dozen new lessons, with more unique styling. There's now a vocabulary lesson template for introducing complex vocabulary that needs explaining—such as すみません , です, and the basic greetings and when they should be used. There's also a new introduction page (planning to turn into a template) for introducing a topic without teaching any material, as well as a 3-column Japanese numbers reference page.

I've also added and populated 168 rows of L1-2 vocab in my Supabase vocab table, and I've integrated it with Google Sheets so I can input data there and I can sync the changes to Postgres. That's using an addon called Coefficient.
I've done a lot of thinking about how I want to search through my vocab list to correct for any unrecognized words generated by Open AI (yet to be implemented). The table is designed to reflect that. Here's an overview of my intentions:
1. OpenAI generates native Japanese with whatever kana is appropriate.
2. Use kuroshiro to convert the text to hiragana and compare it to a sorted list of hiragana in my database.
3. Do a binary search on the hiragana and match the converted word to the hiragana in the database.
4. If a match is found, all is well. If not, respond to OpenAI accordingly to regenerate.
There are a TON of nuances in doing all that, but that's the  basic idea.

I've found a really really good TTS model for Japanese called Voicepeak that I'm definetly planning on integrating in the future as well.