Added Vocab Learn Feature—Multiple-choice and written questions with intelligent shuffling and card cycles.
Added a learn vocab page where users have to answer each question twice: First as multiple-choice, then the card gets pushed back 10 questions and will require a written response instead. If the user gets it correct, it's removed and the next term is added to the cycle of 10. Else, it's put back as multiple choice. Every 7 questions, the user sees a screen that displays the last 7 questions they looked at. This helps the learning feel much less monotonous since a cycle of 10 cards reviewed in sets of 7 feels like it's always changing and unpredictable (helpful when learning).

It also incorporates intelligent, multi-stage shuffling of terms in indexes near the term for multiple-choice questions.
In essence, it's keeping a range of roughly 10-15 entries from the array based around the index of the term that's matched (the index is based on its position in the original data, not the 10 cycled deck terms), and shuffling that range (in a duplicate array), and then selecting the first 3 entries from that shuffled range. Then it adds the term entry (the correct answer) if it's not already in the random entries, and then adds more entries from the array in the shuffled range if randomEntries still has less than 4 entries. Finally, it shuffles the 4 entries.
You can optionally specify to shuffle the original array as well if you want the terms to appear in a random order (not the same order as the json data it pulls from).

Here's a more detailed breakdown of how the shuffle range works:
It converts the JSON object to an array of { key, value } objects
It creates a duplicate array
It gets the index of the term entry
It gets the term entry using the index
It shuffles the duplicate array within the bounds of the lowest range to the highest range (based around the index of the term entry)
It selects the first 3 entries from the shuffled range in the duplicate array
If the term entry exists and is not already in the random entries, it adds it
If randomEntries still has less than 4 entries, it adds more from the array in the shuffled range
It shuffles the random entries