Vocab-learn: Added progress-bar + review page styling. In-lesson practice: multiple-choice highlighting.
The user is given 7 questions at a time during vocab-learn. Added a progress bar that resets after 7 questions.

Some lessons have practice questions within them. Users can now hover to enlarge the text, select it, and it will change the text color to red if incorrect. The text-background will change to green if it's selected and is the correct answer.

Also split the learn-numbers vocab pages into two, (0-10 and 11-100) with 30 additional numbers to practice.