Refactored learn-vocab with sane values for datasets, added mnemonics per-question via dropdown.
Now reads json data items with named properties (hiragana, english, mnemonics), not just string arrays with arbitrarily defined meanings for each index.
In the same light, I added the ability for users to reveal the correct answer (marking it wrong) and it will give the correct answer with the question mnemonic to help them remember it for next time.

*Currently, the spoiler does not auto-hide when the user moves on to the next question. This will definitely be fixed soon.