Lesson pages, learn-vocab-lesson-x, and learn-vocab-lesson-x-kanji pages pull all data from one JSON file per lesson.
Code is way more dry. Instead of writing different variations of the same data in multiple JSON files, it's now all stored in 1 file. Each page can use new components to interpret and reorganize copies of the data as needed.

The standard for JSON data formatting here is:
"眼鏡を掛ける": {
    "hiragana": ["めがねをかける"],
    "english": ["to put on glasses"],
    "mnemonics": "Sounds like 'may-gah-neh-wo-kakeru.' Imagine putting on glasses to see the month of May clearly."
  },
It has a kanji key (or hiragana/katakana if no kanji), with hiragana, english, and mnemonic properties.

The two new components are ConvertToEnglishKey (used so users can look at english keys and write the hiragana as answers in learn-vocab), and VocabDataMapped (maps from data passed as a prop to produce multiple VocabCards in a single line for lesson pages).

In the future, I'd like to make a 3rd component: KanjiFilter for learn-vocab-lesson-x-kanji pages, which filters out all keys that don't contain kanji. Hopefully, I can implement that soon.